Extract TestListener implementation from UnitTestCollectorPlugin
Extracted TestListener implementation to a new class UnitTestListener from UnitTestCollectorPlugin for better separation of concerns.
